Use latest version of hugo
Dependency locking currently doesn't occur in
this repository. We do this for our other
repositories, however I have yet to determine
how to do this for Hugo/Go modules.

Because this isn't a priority, we are setting the
version to use as 'latest' to ensure we take
advantage of bug fixes. All deployments need
to be checked to ensure generation completes
successfully.